pyproject.toml
[project]
name = "forge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.3.5",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "ipykernel>=7.1.0",
    "jupyter-client>=8.8.0",
    "nbformat>=5.10.4",
    "openai>=2.15.0",
    "pytest-asyncio>=1.3.0",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.40.0",
]

[project.scripts]
forge = "forge.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
